Kwami Graphics Update (#6)
* Updated Graphics w/ Kwami
Added the CSS rendered logo to the navigation bar, created CSS
files for the logo and the gameroom. Integrated the Kwami CF2
Demo graphics into the JavaScript, with modifications to use
PIXI as the drawing engine. All JavaScript files updated to use
a variation of the WikiMedia whitespace conventions.

* Updated URL in JavaScript
Changed the URL in the JavaScript files to be dynamic for each
server. Static URLs will prevent being able to easily fork &
push to Heroku, so this is changed in this commit.

* Removed Comments

Deleted the HTML comments from the gameroom template

* Updated cellSize & ID_COLORS

Changed cellSize to 32px for easy sprite compatibility and changed ID_COLORS to be initialized inline rather than dynamically.

* Removed Dynamic Logo Generation

Amended the gamerender.js to remove the logo generation from the initialization. Still need to generate a PNG for the logo to remove the <div> navbar.

* Uploaded Logo

Used Pixilart tool to convert the logo from CSS to PNG. Still need to update the HTML/CSS templates to use the new logo.

* Updated logo.css

Removed the CSS styling for the pixel art logo and replaced it with a background image.

* Update navbar

Removed the <div> based pixel art logo.

* Updated logo size

Changed the background-size attribute of the #colorfight-logo div.

* Updated ID_COLORS

Changed the length of ID_COLORS to be easier to read

* Removed #header from gameroom.css

Deleted artifact from CF2Demo

* Changed to absolute