Added selected cell highlighting. Fixed no border animation.
Restructured cell containers to create objects at init instead of on every
iteration. This makes it easier to manipulate the how the cells draw since
they persist. It appears that we are moving toward a model where we have
data structures that mirror the graphical layout.

Changed the cell per-turn draw to be unconditional. It simplifies the drawing
logic and has no observable performance impact on my side.