Added selected cell highlighting. Fixed no border animation. (#17)
* Added selected cell highlighting. Fixed no border animation.

Restructured cell containers to create objects at init instead of on every
iteration. This makes it easier to manipulate the how the cells draw since
they persist. It appears that we are moving toward a model where we have
data structures that mirror the graphical layout.

Changed the cell per-turn draw to be unconditional. It simplifies the drawing
logic and has no observable performance impact on my side.

* Re-added conditional per-turn draw. Added button to clear selected highlight.

We could make it so the button clears box data, but that requires determining
how to set the defaults in a reasonable way. We could also just make it reset
to [0, 0] or not draw the selected cell at all, but that is somewhat jarring.